+++ source/Api/Configuration/Authentication/Errors/NoAuthenticationHeaderSet.cs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.Infrastructure.Configuration.Authentication.Errors
+namespace Energinet.DataHub.EDI.Api.Configuration.Authentication.Errors
 {
-    public class AuthenticationHeaderIsNotBearerToken : AuthenticationError
+    public class NoAuthenticationHeaderSet : AuthenticationError
     {
-        public AuthenticationHeaderIsNotBearerToken()
-        : base("The value defined in authorization header is not start with 'bearer'.")
+        public NoAuthenticationHeaderSet()
+        : base("No authorization header is set.")
         {
         }
     }
