+++ source/Api/Configuration/Authentication/Errors/TokenValidationFailed.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.Infrastructure.Configuration.Authentication.Errors
+namespace Energinet.DataHub.EDI.Api.Configuration.Authentication.Errors
 {
     public class TokenValidationFailed : AuthenticationError
     {
