+++ source/IntegrationTests/Infrastructure/Authentication/MarketActors/JwtTokenParserTests.cs
@@ -16,8 +16,8 @@ using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.Authentication;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.Authentication.Errors;
+using Energinet.DataHub.EDI.Api.Configuration.Authentication;
+using Energinet.DataHub.EDI.Api.Configuration.Authentication.Errors;
 using Microsoft.IdentityModel.Tokens;
 using Xunit;
 using Xunit.Categories;
