+++ source/IntegrationTests/TestBase.cs
@@ -24,6 +24,7 @@ using BuildingBlocks.Application.FeatureFlag;
 using Dapper;
 using Energinet.DataHub.EDI.Api;
 using Energinet.DataHub.EDI.Api.Authentication;
+using Energinet.DataHub.EDI.Api.Configuration.Authentication;
 using Energinet.DataHub.EDI.Api.Configuration.Middleware.Correlation;
 using Energinet.DataHub.EDI.ArchivedMessages.Application.Extensions.DependencyInjection;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Authentication;
@@ -266,6 +267,8 @@ namespace Energinet.DataHub.EDI.IntegrationTests
 
             CompositionRoot.Initialize(_services)
                 .AddSystemClock(new SystemDateTimeProviderStub());
+
+            _services.AddScoped(_ => new JwtTokenParser(JwtTokenParserTests.DisableAllTokenValidations));
             _services.AddScoped(_ =>
             {
                 var correlation = new CorrelationContext();
